Allow using 1.12.x patch versions of dub
The `~>1.12.0` syntax (which is short for `>=1.12.0 <1.13.0`) should be preferred
as it allows bug fixes from dub to propagate to dls, while preventing breaking
changes. For this to work, the exact dub version should be removed from
`dub.selections.json`, as this file overrides any flexibility in
`dub.json` / `dub.sdl`.

See the pull request which merged this commit for more details.